requirements.txt
# Core data science
numpy>=1.21
pandas>=1.3
scipy>=1.7
scikit-learn>=1.0
matplotlib>=3.4
seaborn>=0.11

# NLP and embeddings
sentence-transformers>=2.2
faiss-cpu>=1.7   # or faiss-gpu if you have CUDA
torch>=1.10      # required by sentence-transformers

# Deep learning
tensorflow>=2.11
transformers>=4.25
tf-keras>=2.11   # for TF2 / Keras3 compatibility

# Utilities
contractions>=0.1.72
wordcloud>=1.8
tqdm>=4.64
joblib>=1.1
pyyaml>=6.0
dask[distributed]>=2022.12
mlflow>=2.3

# Data loading
tensorflow-datasets>=4.8

bokeh>=3.1.0
vaderSentiment